@@ -4,8 +4,8 @@ version: 0.0.1
 homepage:
 
 environment:
-  sdk: ">=3.3.0 <4.0.0"
-  flutter: ">=3.24.0"
+  sdk: ">=3.7.0 <4.0.0"
+  flutter: ">=3.29.0"
 publish_to: 'none'
 
 dependencies:
@@ -15,6 +15,6 @@ dependencies:
   json_annotation: ^4.9.0
 
 dev_dependencies:
-  flutter_lints: ^4.0.0
-  build_runner: ^2.4.11
-  json_serializable: ^6.8.0
+  flutter_lints: ^5.0.0
+  build_runner: ^2.4.15
+  json_serializable: ^6.9.4
