@@ -4,8 +4,8 @@ version: 0.0.1
 homepage:
 
 environment:
-  sdk: ">=3.3.0 <4.0.0"
-  flutter: ">=3.24.0"
+  sdk: ">=3.7.0 <4.0.0"
+  flutter: ">=3.29.0"
 
 publish_to: none
 
@@ -17,15 +17,13 @@ dependencies:
   content_suppliers_api:
     path: ../content_suppliers_api
   freezed_annotation: ^2.4.4
-  collection: ^1.19.0
+  collection: ^1.19.1
   cached_network_image: ^3.4.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^4.0.0
+  flutter_lints: ^5.0.0
   freezed: ^2.5.7
-  integration_test:
-    sdk: flutter
-  build_runner: ^2.4.13
-  test: ^1.25.7
+  build_runner: ^2.4.15
+  test: 
